+++ client-side/src/components/Dashboard/ContentDashboard.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function ContentDashboard() {
+  return (
+    <div>ContentDashboard</div>
+  )
+}
+
+export default ContentDashboard
